Day 7 Part 2
- extracted functions for parsing the hand type and its power
- implemented the normal and joker parsings

+++ Year 2023/src/main/java/com/github/silviuburceadev/aoc/camelcard/Game.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public record Game(List<PlayingHand> playingHands) {
 
-    public int winnings() {
+    public long winnings() {
         List<PlayingHand> copy = new ArrayList<>(playingHands);
         Collections.sort(copy);
-        int total = 0;
+        long total = 0;
         for (int i = 0; i < copy.size(); i++) {
             total += copy.get(i).bet() * (i + 1);
         }
